+++ examples/air-quality-sensor-app/silabs/BUILD.gn
@@ -124,8 +124,6 @@ silabs_executable("air_quality_sensor_app") {
   sources = [
     "${chip_root}/examples/air-quality-sensor-app/air-quality-sensor-common/src/air-quality-sensor-manager.cpp",
     "src/AppTask.cpp",
-    "src/DataModelCallbacks.cpp",
-    "src/SensorManager.cpp",
   ]
 
   #TODO: This is a placeholder for the actual implementation
