+++ examples/air-quality-sensor-app/silabs/src/AppTask.cpp
@@ -34,20 +34,32 @@
 #endif // QR_CODE_ENABLED
 #endif // DISPLAY_ENABLED
 
+#include <AirQualityConfig.h>
 #include <air-quality-sensor-manager.h>
 #include <app-common/zap-generated/attributes/Accessors.h>
 #include <app-common/zap-generated/callback.h>
 #include <app-common/zap-generated/cluster-objects.h>
 #include <app-common/zap-generated/ids/Attributes.h>
+#include <app-common/zap-generated/ids/Clusters.h>
+#include <app/ConcreteAttributePath.h>
 #include <app/server/Server.h>
 #include <app/util/attribute-storage.h>
 #include <assert.h>
 #include <lib/support/CodeUtils.h>
 #include <platform/CHIPDeviceLayer.h>
+#include <platform/PlatformError.h>
 #include <platform/silabs/platformAbstraction/SilabsPlatform.h>
 #include <setup_payload/OnboardingCodesUtil.h>
 #include <setup_payload/QRCodeSetupPayloadGenerator.h>
 #include <setup_payload/SetupPayload.h>
+
+#ifdef USE_AIR_QUALITY_SENSOR
+#include "AirQualitySensor.h"
+#endif
+
+#ifdef SL_MATTER_ENABLE_AWS
+#include "MatterAwsControl.h"
+#endif // SL_MATTER_ENABLE_AWS
 
 /**********************************************************
  * Defines and Constants
@@ -62,11 +74,150 @@ using namespace chip::DeviceLayer;
 using namespace chip::app::Clusters::AirQuality;
 using namespace chip::app::Clusters;
 
+namespace {
+constexpr uint16_t kSensorTImerPeriodMs = 30000; // 30s timer period
+
+#ifndef USE_AIR_QUALITY_SENSOR
+constexpr uint16_t kSimulatedReadingFrequency =
+    (60000 / kSensorTImerPeriodMs); // for every two timer cycles, a simulated sensor update is triggered.
+int32_t mSimulatedAirQuality[] = { 5, 55, 105, 155, 205, 255, 305, 355, 400 };
+#endif
+
+} // namespace
+
+/**
+ * @brief Classifies the air quality based on a given sensor value.
+ *
+ * This function compares the input value against predefined thresholds
+ * defined in the AirQualityConfig.h file. The thresholds are used
+ * to classify the air quality into categories defined by the AirQualityEnum.
+ * The thresholds are defined in the SensorThresholds enum.
+ *
+ * @param value The sensor value used to classify air quality.
+ * @return AirQualityEnum The classified air quality category.
+ */
+AirQualityEnum classifyAirQuality(int32_t value)
+{
+    if (value < MIN_THRESHOLD)
+    {
+        return AirQualityEnum::kUnknown;
+    }
+    else if (value < GOOD_THRESHOLD)
+    {
+        return AirQualityEnum::kGood;
+    }
+    else if (value < FAIR_THRESHOLD)
+    {
+        return AirQualityEnum::kFair;
+    }
+    else if (value < MODERATE_THRESHOLD)
+    {
+        return AirQualityEnum::kModerate;
+    }
+    else if (value < POOR_THRESHOLD)
+    {
+        return AirQualityEnum::kPoor;
+    }
+    else if (value < VERY_POOR_THRESHOLD)
+    {
+        return AirQualityEnum::kVeryPoor;
+    }
+    else
+    {
+        return AirQualityEnum::kExtremelyPoor;
+    }
+}
+
+void InitAirQualitySensorManager(intptr_t arg)
+{
+    (void) arg;
+    AirQualitySensorManager::InitInstance();
+}
+
+void writeAirQualityToAttribute(intptr_t context)
+{
+    int32_t * air_quality_ptr = reinterpret_cast<int32_t *>(context);
+    AirQualitySensorManager::GetInstance()->OnAirQualityChangeHandler(classifyAirQuality(*air_quality_ptr));
+    ChipLogDetail(AppServer, "RAW AirQuality value: %ld and corresponding Enum value : %d", *air_quality_ptr,
+                  chip::to_underlying(AirQualitySensorManager::GetInstance()->GetAirQuality()));
+    AppTask::GetAppTask().UpdateAirQualitySensorUI();
+    delete air_quality_ptr;
+}
+
 /**********************************************************
  * AppTask Definitions
  *********************************************************/
 
 AppTask AppTask::sAppTask;
+
+CHIP_ERROR AppTask::InitAirQualitySensor()
+{
+    TEMPORARY_RETURN_IGNORED DeviceLayer::PlatformMgr().ScheduleWork(InitAirQualitySensorManager);
+    // Create cmsisos sw timer for air quality sensor timer.
+    mSensorTimer = osTimerNew(SensorTimerEventHandler, osTimerPeriodic, nullptr, nullptr);
+    if (mSensorTimer == NULL)
+    {
+        ChipLogDetail(AppServer, "mSensorTimer timer create failed");
+        return APP_ERROR_CREATE_TIMER_FAILED;
+    }
+
+#ifdef USE_AIR_QUALITY_SENSOR
+    sl_status_t status = AirQualitySensor::Init();
+    if (status != SL_STATUS_OK)
+    {
+        ChipLogError(AppServer, "Failed to Init Sensor with error code: %lx", status);
+        return MATTER_PLATFORM_ERROR(status);
+    }
+#endif
+    // Update Air Quality immediatly at bootup
+    SensorTimerEventHandler(nullptr);
+    // Trigger periodic update
+    uint32_t delayTicks = ((uint64_t) osKernelGetTickFreq() * kSensorTImerPeriodMs) / 1000;
+
+    // Starts or restarts the function timer
+    if (osTimerStart(mSensorTimer, delayTicks))
+    {
+        ChipLogDetail(AppServer, "mSensor Timer start() failed");
+        appError(APP_ERROR_START_TIMER_FAILED);
+    }
+    return CHIP_NO_ERROR;
+}
+
+void AppTask::SensorTimerEventHandler(void * arg)
+{
+    int32_t air_quality;
+#ifdef USE_AIR_QUALITY_SENSOR
+    if (SL_STATUS_OK != AirQualitySensor::GetAirQuality(air_quality))
+    {
+        ChipLogDetail(AppServer, "Failed to read Air Quality !!!");
+        return;
+    }
+#else
+    // Initialize static variables to keep track of the current index and repetition count
+    static uint8_t nbOfRepetition = 0;
+    static uint8_t simulatedIndex = 0;
+
+    // Ensure the simulatedIndex wraps around the array size to avoid out-of-bounds access
+    simulatedIndex = simulatedIndex % MATTER_ARRAY_SIZE(mSimulatedAirQuality);
+    // Retrieve the current air quality value from the simulated data array using the simulatedIndex
+    air_quality = mSimulatedAirQuality[simulatedIndex];
+
+    // Increment the repetition count
+    nbOfRepetition++;
+    // Check if the number of repetitions has reached the threshold to simulate a new reading
+    if (nbOfRepetition >= kSimulatedReadingFrequency)
+    {
+        // Move to the next index for the next simulated reading
+        simulatedIndex++;
+        // Reset the repetition count
+        nbOfRepetition = 0;
+    }
+#endif // USE_AIR_QUALITY_SENSOR
+    // create pointer for the int32_t air_quality
+    int32_t * air_quality_ptr = new int32_t(air_quality);
+    TEMPORARY_RETURN_IGNORED DeviceLayer::PlatformMgr().ScheduleWork(writeAirQualityToAttribute,
+                                                                     reinterpret_cast<intptr_t>(air_quality_ptr));
+}
 
 CHIP_ERROR AppTask::AppInit()
 {
@@ -76,10 +227,10 @@ CHIP_ERROR AppTask::AppInit()
     GetLCD().SetCustomUI(AirQualitySensorUI::DrawUI);
 #endif
 
-    err = SensorManager::SensorMgr().Init();
+    err = InitAirQualitySensor();
     if (err != CHIP_NO_ERROR)
     {
-        ChipLogDetail(AppServer, "SensorMgr::Init() failed");
+        ChipLogDetail(AppServer, "InitAirQualitySensor() failed");
         appError(err);
     }
 
@@ -144,4 +295,24 @@ void AppTask::ButtonEventHandler(uint8_t button, uint8_t btnAction)
         aEvent.Handler = BaseApplication::ButtonHandler;
         sAppTask.PostEvent(&aEvent);
     }
+}
+
+void AppTask::DMPostAttributeChangeCallback(const ConcreteAttributePath & attributePath, uint8_t type, uint16_t size,
+                                            uint8_t * value)
+{
+    ClusterId clusterId                      = attributePath.mClusterId;
+    [[maybe_unused]] AttributeId attributeId = attributePath.mAttributeId;
+    ChipLogDetail(Zcl, "Cluster callback: " ChipLogFormatMEI, ChipLogValueMEI(clusterId));
+
+    if (clusterId == Identify::Id)
+    {
+        ChipLogProgress(Zcl, "Identify attribute ID: " ChipLogFormatMEI " Type: %u Value: %u, length %u",
+                        ChipLogValueMEI(attributeId), type, *value, size);
+    }
+}
+
+void MatterPostAttributeChangeCallback(const chip::app::ConcreteAttributePath & attributePath, uint8_t type, uint16_t size,
+                                       uint8_t * value)
+{
+    AppTask::GetAppTask().DMPostAttributeChangeCallback(attributePath, type, size, value);
 }
